Added menu functionality with buttons. Future updates will modify the appearance of inventory and panels. Music and sound settings implemented. A pause panel triggered by the Escape key is planned.
A menu has been added with fully functional buttons. The appearance of inventory and other panels, including slots, will be updated in future updates. Music has been integrated into the menu, and sound settings are now available. In the future, the ability to pause the game via the Escape key, opening a pause menu, will be implemented.